+++ src/main/java/com/mongodb/kafka/connect/util/ClassHelper.java
@@ -66,6 +66,12 @@ public final class ClassHelper {
     } catch (ClassNotFoundException e) {
       throw new ConnectConfigException(
           configKey, className, format("Class not found: %s", e.getMessage()));
+    } catch (NoSuchMethodException e) {
+      throw new ConnectConfigException(
+          configKey,
+          className,
+          format(
+              "Class could not be initialized, no public default constructor: %s", e.getMessage()));
     } catch (Exception e) {
       if (e.getCause() instanceof ConnectConfigException) {
         throw (ConnectConfigException) e.getCause();